Optimizacion promociones

+++ app/Services/Reservation/Promotion/PromotionService.php
@@ -27,12 +27,16 @@ class PromotionService
 
     public function SavePromotion($data, int $user_id)
     {
+
         //construir el array de redimensiones
         $dimensionsImg = ManageFilesHelper::GetDimensionsPromotionImg();
 
         //crear instancia de image service
         $imageService = new ImageService();
+
         try {
+            if(!empty($data['image'])){
+
             $imageService->ResizeImage(@$data['image_fullname'], @$data['image'], 'promotions', $dimensionsImg, @$data['cropper']['areaCoords'], @$data['cropper']['canvasWidth']);
             $data['user_id'] = '1';
             $url = $this->API_PROMO_URL . '/es/microsites/1/promotions';
@@ -47,12 +51,21 @@ class PromotionService
                 ManageFilesHelper::RemoveImagesFromDimensions($imageService, 'promotions', $dimensionsImg, $data['image']);
             }
             return $response;
+
+            }else{
+                $url = $this->API_PROMO_URL . '/es/microsites/1/promotions';
+                $response = ApiRequestsHelper::SendRequest('POST', $url, null, $data);
+                //dd($response);
+               return $response;
+            }
+
         } catch (NotReadableException $e) {
             abort(400, trans('messages.image_not_readable'));
         } catch (\Exception $e) {
         	//var_dump($e->getMessage());exit;
             abort(500, trans('error.500'));
         }
+
     }
 
     public function UpdatePromotion($data, $micro_id, $promotion_id, int $user_id)
@@ -66,33 +79,41 @@ class PromotionService
 
         $promotion = $resp['data'];
         try {
-            if ($data['image'] == $promotion['image']) {
-                $imgFullname = '/files/promotions/image/800x800/' . @$data['image'];
-                $imageService->ResizeImage($imgFullname, @$data['image'], 'promotions', $dimensionsImg, @$data['cropper']['areaCoords'], @$data['cropper']['canvasWidth'], false);
-            } else {
-                $imageService->ResizeImage(@$data['image_fullname'], @$data['image'], 'promotions', $dimensionsImg, @$data['cropper']['areaCoords'], @$data['cropper']['canvasWidth']);
-            }
-
-            //$data['user_id'] = $user_id;
-            $data['user_id'] = '1';
-            $url = $this->API_PROMO_URL . '/es/microsites/1/promotions/' . $promotion_id;
-            //Se envía la solicitud al api
-            //echo $promotion_id;
+            //dd($data);
             //die();
-            $response = ApiRequestsHelper::SendRequest('PUT', $url, null, $data);
-            dd($response);
-            //exit();
+            if(!empty($data['image'])){
 
-            if ($response['success']) {
-                //Se borran las imagenes de las carpetas temporales
-                if (isset($data['image_fullname'])) {
-                    $imageService->RemoveImage($data['image_fullname']);
+                if ($data['image'] == $promotion['image']) {
+                    $imgFullname = '/files/promotions/image/800x800/' . @$data['image'];
+                    $imageService->ResizeImage($imgFullname, @$data['image'], 'promotions', $dimensionsImg, @$data['cropper']['areaCoords'], @$data['cropper']['canvasWidth'], false);
+                } else {
+                    $imageService->ResizeImage(@$data['image_fullname'], @$data['image'], 'promotions', $dimensionsImg, @$data['cropper']['areaCoords'], @$data['cropper']['canvasWidth']);
                 }
-            } else {
-                //Se borran las imagenes de las carpetas de redimension.
-                ManageFilesHelper::RemoveImagesFromDimensions($imageService, 'promotions', $dimensionsImg, $data['image'], false);
+
+                //$data['user_id'] = $user_id;
+                $data['user_id'] = '1';
+                $url = $this->API_PROMO_URL . '/es/microsites/1/promotions/' . $promotion_id;
+                //Se envía la solicitud al api
+                $response = ApiRequestsHelper::SendRequest('PUT', $url, null, $data);
+
+                if ($response['success']) {
+                    //Se borran las imagenes de las carpetas temporales
+                    if (isset($data['image_fullname'])) {
+                        $imageService->RemoveImage($data['image_fullname']);
+                    }
+                } else {
+                    //Se borran las imagenes de las carpetas de redimension.
+                    ManageFilesHelper::RemoveImagesFromDimensions($imageService, 'promotions', $dimensionsImg, $data['image'], false);
+                }
+                return $response;
+
+            }else{
+                $url = $this->API_PROMO_URL . '/es/microsites/1/promotions/' . $promotion_id;
+                $response = ApiRequestsHelper::SendRequest('PUT', $url, null, $data);
+                //dd($response);
+               return $response;
             }
-            return $response;
+
         } catch (NotReadableException $e) {
             abort(400, trans('messages.image_not_readable'));
         } catch (\Exception $e) {
